refactor: ViewController에서 ChapterListView 분리 및 적용
- 기존 챕터 UI를 ChapterListView로 분리하여 역할 명확화
- ViewController에서 chapterStackView 및 chapter label 생성/관리 로직 제거
- ViewController는 ChapterListView에 데이터만 전달하도록 개선
- ChapterListView에서 타이틀과 챕터 리스트 동적 생성 및 관리
